slipt old Summary test class into Node and Namespace
